"ads2_m6_a1_1: 2/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2-assignments/m6/ADS-2Week-1/Solution.java
@@ -6,7 +6,7 @@ class PageRank {
 		this.digr = df;
 		pray = new double[digr.V()];
 		for (int i = 0; i < pray.length; i++) {
-			pray[i] = 1.0 / digr.V();
+			pray[i] = 1.0 / (double)digr.V();
 		}
 	}
